+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Next version
 
 - Fixed crash on nested values on the `#enabled?` method. See [#3](https://github.com/mssola/cconfig/issues/3).
+- The `prefix` argument from the constructor now takes a default
+  (`cconfig`). This will make things more predictable I hope.
 
 ## 1.1.0